src/requirements.txt
browser-cookie3==0.20.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
idna==3.11
jeepney==0.9.0
lz4==4.4.5
pycryptodomex==3.23.0
PyYAML==6.0.3
requests==2.33.1
tqdm==4.67.3
urllib3==2.6.3
